Regras de connect e disconnect mudadas

+++ dropboxUtil.h
@@ -50,7 +50,7 @@ struct FileInfo {
 struct Client {
     std::string user_id;
     bool is_logged;
-    long connected_devices[MAX_DEVICES];
+    int connected_devices[MAX_DEVICES];
     int sockets[MAX_DEVICES];
     std::vector<FileInfo> files;
